+++ spec/ultra_settings/uninitialized_runtime_settings_spec.rb
@@ -2,7 +2,7 @@
 
 require_relative "../spec_helper"
 
-describe UltraSettings::UninitializedRuntimeSettings do
+RSpec.describe UltraSettings::UninitializedRuntimeSettings do
   it "raises an error when trying to access a value" do
     expect { UltraSettings::UninitializedRuntimeSettings["foo"] }.to raise_error(UltraSettings::UninitializedRuntimeSettings::Error)
   end